Tons of queued changes from the last six months of my private fork. Some good UI cleanup and added the Activations logic.

+++ order.php
@@ -12,11 +12,21 @@
 	if(isset($_GET['act']) && $_GET['act'] == 'download')
 		$o->downloadLicense();
 
+	if(isset($_GET['act']) && $_GET['act'] == 'upgrade')
+	{
+		$upgraded_order = $o->upgradeLicense();
+		redirect('order.php?id=' . $upgraded_order->id);
+		exit;
+	}
+	
+
 	if(isset($_GET['act']) && $_GET['act'] == 'delete')
 	{
 		$o->delete();
 		redirect('orders.php');
 	}
+
+	$app = new Application($o->app_id);
 ?>
 <?PHP include('inc/header.inc.php'); ?>
 
@@ -64,8 +74,20 @@
 					</div>
 					<div class="bd">
 						<ul class="biglist">
+							<?PHP if($app->upgrade_app_id > 0) : ?>
+							<li><a href="order.php?id=<?PHP echo $o->id; ?>&amp;act=upgrade" id="upgrade">Upgrade Order</a></li>
+							<?PHP endif; ?>
 							<li><a href="order.php?id=<?PHP echo $o->id; ?>&amp;act=delete" id="delete">Delete Order</a></li>
 						</ul>
+					</div>
+				</div>
+
+				<div class="block">
+					<div class="hd">
+						<h3>Activations</h3>
+					</div>
+					<div class="bd">
+						<a href="activations.php?q=<?PHP echo $o->payer_email; ?>">Activated <?PHP echo $o->activationCount(); ?> times</a>
 					</div>
 				</div>
 
@@ -74,7 +96,11 @@
 						<h3>Cut &amp; Paste License</h3>
 					</div>
 					<div class="bd">
+						<?PHP if($app->license_type == 'ap') : ?>
+						<textarea style="width:100%;"><?PHP echo $o->license; ?></textarea>
+						<?PHP else : ?>
 						<textarea style="width:100%;"><?PHP echo "Email: {$o->payer_email}\nReg Key: {$o->license}"; ?></textarea>
+						<?PHP endif; ?>
 					</div>
 				</div>
